fix: move sdk outputs into build directory and do not hide fact that this is not an ESM module so this only worked because we bundled everything into the same directory
