+++ plugins/conftest-container/index.ts
@@ -8,7 +8,7 @@
 
 import { createGardenPlugin } from "@garden-io/sdk"
 import { ConftestProvider } from "@garden-io/garden-conftest"
-import { dedent } from "@garden-io/sdk/util/string"
+import { dedent } from "@garden-io/sdk/build/src/util/string"
 
 /**
  * Auto-generates a conftest module for each container module in your project
