+++ plugins/conftest-kubernetes/test/conftest-kubernetes.ts
@@ -12,8 +12,8 @@ import { join } from "path"
 
 import { gardenPlugin } from ".."
 import { gardenPlugin as conftestPlugin } from "@garden-io/garden-conftest"
-import { dedent } from "@garden-io/sdk/util/string"
-import { makeTestGarden } from "@garden-io/sdk/testing"
+import { dedent } from "@garden-io/sdk/build/src/util/string"
+import { makeTestGarden } from "@garden-io/sdk/build/src/testing"
 
 import { TestTask } from "@garden-io/core/build/src/tasks/test"
