+++ plugins/conftest/test/conftest.ts
@@ -10,11 +10,11 @@ import { expect } from "chai"
 import stripAnsi from "strip-ansi"
 import { join } from "path"
 
-import { dedent } from "@garden-io/sdk/util/string"
-import { defaultNamespace } from "@garden-io/sdk/constants"
+import { dedent } from "@garden-io/sdk/build/src/util/string"
+import { defaultNamespace } from "@garden-io/sdk/build/src/constants"
 import { gardenPlugin } from ".."
-import { ProjectConfig } from "@garden-io/sdk/types"
-import { makeTestGarden } from "@garden-io/sdk/testing"
+import { ProjectConfig } from "@garden-io/sdk/build/src/types"
+import { makeTestGarden } from "@garden-io/sdk/build/src/testing"
 
 import { TestTask } from "@garden-io/core/build/src/tasks/test"
 import { defaultDotIgnoreFile } from "@garden-io/core/build/src/util/fs"
