+++ plugins/jib/mavend.ts
@@ -7,7 +7,7 @@
  */
 
 import AsyncLock from "async-lock"
-import { PluginContext, PluginToolSpec } from "@garden-io/sdk/types"
+import { PluginContext, PluginToolSpec } from "@garden-io/sdk/build/src/types"
 import { find } from "lodash"
 import { PluginError } from "@garden-io/core/build/src/exceptions"
 import { BuildToolParams, runBuildTool, verifyBinaryPath, VerifyBinaryParams } from "./build-tool-base"
