+++ plugins/terraform/commands.ts
@@ -9,9 +9,9 @@
 import chalk from "chalk"
 import { terraform } from "./cli"
 import { TerraformProvider } from "./provider"
-import { ConfigurationError, ParameterError } from "@garden-io/sdk/exceptions"
+import { ConfigurationError, ParameterError } from "@garden-io/sdk/build/src/exceptions"
 import { prepareVariables, setWorkspace, tfValidate } from "./helpers"
-import { ConfigGraph, PluginCommand, PluginCommandParams } from "@garden-io/sdk/types"
+import { ConfigGraph, PluginCommand, PluginCommandParams } from "@garden-io/sdk/build/src/types"
 import { join } from "path"
 import { remove } from "fs-extra"
 import { getProviderStatusCachePath } from "@garden-io/core/build/src/tasks/resolve-provider"
