+++ plugins/terraform/index.ts
@@ -11,11 +11,11 @@ import { omit } from "lodash"
 import { pathExists } from "fs-extra"
 import { createGardenPlugin } from "@garden-io/sdk"
 import { cleanupEnvironment, getEnvironmentStatus, prepareEnvironment } from "./init"
-import { dedent } from "@garden-io/sdk/util/string"
+import { dedent } from "@garden-io/sdk/build/src/util/string"
 import { defaultTerraformVersion, terraformCliSpecs } from "./cli"
-import { ConfigurationError } from "@garden-io/sdk/exceptions"
+import { ConfigurationError } from "@garden-io/sdk/build/src/exceptions"
 import { configureTerraformModule, TerraformModule, terraformModuleSchema } from "./module"
-import { docsBaseUrl } from "@garden-io/sdk/constants"
+import { docsBaseUrl } from "@garden-io/sdk/build/src/constants"
 import { getTerraformCommands } from "./commands"
 import { TerraformDeployConfig, terraformDeployOutputsSchema, terraformDeploySchema } from "./action"
 import { deleteTerraformModule, deployTerraform, getTerraformStatus } from "./handlers"
