+++ plugins/terraform/module.ts
@@ -9,8 +9,8 @@
 import { join } from "path"
 import { pathExists } from "fs-extra"
 import { joi } from "@garden-io/core/build/src/config/common"
-import { GardenModule, ModuleActionHandlers, PluginContext } from "@garden-io/sdk/types"
-import { ConfigurationError } from "@garden-io/sdk/exceptions"
+import { GardenModule, ModuleActionHandlers, PluginContext } from "@garden-io/sdk/build/src/types"
+import { ConfigurationError } from "@garden-io/sdk/build/src/exceptions"
 import { dependenciesSchema } from "@garden-io/core/build/src/config/service"
 import { TerraformBaseSpec } from "./helpers"
 import { TerraformProvider, TerraformProviderConfig } from "./provider"
