+++ plugins/terraform/test/common.ts
@@ -10,8 +10,8 @@ import { join } from "path"
 import { pathExists, remove } from "fs-extra"
 import { gardenPlugin } from ".."
 import { TerraformProvider } from "../provider"
-import { makeTestGarden, TestGarden } from "@garden-io/sdk/testing"
-import { Log, PluginContext } from "@garden-io/sdk/types"
+import { makeTestGarden, TestGarden } from "@garden-io/sdk/build/src/testing"
+import { Log, PluginContext } from "@garden-io/sdk/build/src/types"
 import { getWorkspaces, setWorkspace } from "../helpers"
 import { expect } from "chai"
 import { defaultTerraformVersion, terraform } from "../cli"
