+++ plugins/terraform/test/terraform.ts
@@ -11,10 +11,10 @@ import { join } from "path"
 import { expect } from "chai"
 import { pathExists, readFile, remove } from "fs-extra"
 
-import { getRootLogMessages, makeTestGarden, TestGarden } from "@garden-io/sdk/testing"
+import { getRootLogMessages, makeTestGarden, TestGarden } from "@garden-io/sdk/build/src/testing"
 import { findByName } from "@garden-io/core/build/src/util/util"
 import { getTerraformCommands } from "../commands"
-import { ConfigGraph, LogLevel } from "@garden-io/sdk/types"
+import { ConfigGraph, LogLevel } from "@garden-io/sdk/build/src/types"
 import { gardenPlugin } from ".."
 import { TerraformProvider } from "../provider"
 import { DeployTask } from "@garden-io/core/build/src/tasks/deploy"
