+++ sdk/src/exceptions.ts
