+++ sdk/src/types.ts
@@ -17,7 +17,7 @@ export { LogLevel } from "@garden-io/core/build/src/logger/logger"
 export { Log, ActionLog } from "@garden-io/core/build/src/logger/log-entry"
 export { PluginContext } from "@garden-io/core/build/src/plugin-context"
 export { ProjectConfig } from "@garden-io/core/build/src/config/project"
-export { PluginToolSpec } from "../core/build/src/plugin/tools"
+export { PluginToolSpec } from "@garden-io/core/build/src/plugin/tools"
 export { GardenPluginSpec as GardenPlugin } from "@garden-io/core/build/src/plugin/plugin"
 export { ConfigGraph } from "@garden-io/core/build/src/graph/config-graph"
 export { PluginCommand, PluginCommandParams } from "@garden-io/core/build/src/plugin/command"
