+++ sdk/src/util/ext-tools.ts
