+++ sdk/src/util/string.ts
